Added DomResultExtensions.cs Added DrawingExtensions.cs Added ElementArea.cs Moved SciterNode from SciterElement.cs to SciterNode.cs Fixed SCITER_FIRE_EVENT in Sciter.cs Just about finished on SciterElementExtensions.cs and SciterElement.cs Correcting SciterXDom.cs Summaries. Added T_RESOURCE and T_ASSET to VALUE_TYPE and SciterValue.cs

+++ Shared/SciterCore.Shared/Enums/ElementState.cs
@@ -3,7 +3,7 @@
 namespace SciterCore
 {
 	[Flags]
-    public enum ElementState : int
+    public enum ElementState : long
     {
 	    None		     = 0,
